refactor: split meeting_facilitator.rs (511 LOC) into sub-modules
Split monolithic src/meeting_facilitator.rs into src/meeting_facilitator/:
- mod.rs: module doc + re-exports (19 lines)
- types.rs: MeetingDecision, ActionItem, MeetingSession, MeetingSessionStatus (76 lines)
- session.rs: validation helpers + session API + tests (255 lines)
- handoff.rs: MeetingHandoff + file I/O (182 lines)

All files under the 400-line target. All pub items re-exported from mod.rs
so downstream imports are unchanged. Fixes E0761 dual-file conflict.
