+++ src/meeting_facilitator/mod.rs
@@ -1,0 +1,19 @@
+//! Interactive meeting mode — structured session with decisions and action items.
+//!
+//! `MeetingSession` captures a running meeting with its topic, decisions made,
+//! and action items assigned. The facilitator stores a durable summary into
+//! cognitive memory when the meeting closes.
+
+mod handoff;
+mod session;
+mod types;
+
+// Re-export all public items so `crate::meeting_facilitator::X` still works.
+pub use handoff::{
+    MEETING_HANDOFF_FILENAME, MeetingHandoff, default_handoff_dir, load_meeting_handoff,
+    mark_handoff_processed_in_place, mark_meeting_handoff_processed, write_meeting_handoff,
+};
+pub use session::{
+    add_note, close_meeting, record_action_item, record_decision, start_meeting,
+};
+pub use types::{ActionItem, MeetingDecision, MeetingSession, MeetingSessionStatus};
